Flake8 verification, fixed all tests and used the new function instead
