+++ dipy/workflows/tests/test_viz.py
@@ -1,15 +1,15 @@
 import os
-import numpy as np
-import numpy.testing as npt
+
+from dipy.io.image import save_nifti
+from dipy.io.stateful_tractogram import Space, StatefulTractogram
+from dipy.io.streamline import save_tractogram
+from dipy.io.utils import create_nifti_header
 from dipy.tracking.streamline import Streamlines
 from dipy.testing.decorators import xvfb_it, use_xvfb
 from dipy.utils.optpkg import optional_package
 from nibabel.tmpdirs import TemporaryDirectory
-from dipy.io.image import save_nifti
-from dipy.io.stateful_tractogram import (create_nifti_header, Space,
-                                         StatefulTractogram)
-from dipy.io.streamline import save_tractogram
-
+import numpy as np
+import numpy.testing as npt
 
 fury, has_fury, setup_module = optional_package('fury')
 
@@ -81,7 +81,7 @@ def test_horizon_flow():
         voxel_sizes = np.array([2, 1.0, 1.0])
         nii_header = create_nifti_header(affine, dimensions, voxel_sizes)
         sft = StatefulTractogram(streamlines, nii_header, space=Space.RASMM)
-        save_tractogram(sft, ftrk)
+        save_tractogram(sft, ftrk, bbox_valid_check=False)
 
         input_files = [ftrk, fimg]
